General working client, small kinks to solve still
