+++ graphql_client/Cargo.toml
@@ -1,68 +1,60 @@
 [package]
-name = "graphql_client"
-version = "0.9.0"
 authors = ["Tom Houlé <tom@tomhoule.com>"]
-description = "Typed GraphQL requests and responses"
-repository = "https://github.com/graphql-rust/graphql-client"
-license = "Apache-2.0 OR MIT"
-keywords = ["graphql",  "api", "web", "webassembly", "wasm"]
 categories = ["network-programming", "web-programming", "wasm"]
+description = "Typed GraphQL requests and responses"
 edition = "2018"
+keywords = ["graphql", "api", "web", "webassembly", "wasm"]
+license = "Apache-2.0 OR MIT"
+name = "graphql_client"
+repository = "https://github.com/graphql-rust/graphql-client"
+version = "0.9.0"
 
 [dependencies]
+anyhow = {version = "1.0", optional = true}
 doc-comment = "^0.3"
-anyhow = { version = "1.0", optional = true }
-thiserror = { version = "1.0", optional = true }
-graphql_query_derive = { path = "../graphql_query_derive", version = "0.9.0" }
+graphql_query_derive = {path = "../graphql_query_derive", version = "0.9.0"}
+reqwest = {version = "0.10.7", optional = true}
+serde = {version = "^1.0.78", features = ["derive"]}
 serde_json = "1.0"
-serde = { version = "^1.0.78", features = ["derive"] }
+thiserror = {version = "1.0", optional = true}
 
 [dependencies.futures]
-version = "^0.1"
 optional = true
+version = "^0.1"
 
 [dependencies.js-sys]
-version = "^0.3"
 optional = true
+version = "^0.3"
 
 [dependencies.log]
-version = "^0.4"
 optional = true
+version = "^0.4"
 
 [dependencies.web-sys]
-version = "^0.3"
-optional = true
 features = [
-    "Headers",
-    "Request",
-    "RequestInit",
-    "Response",
-    "Window",
+  "Headers",
+  "Request",
+  "RequestInit",
+  "Response",
+  "Window",
 ]
+optional = true
+version = "^0.3"
 
 [dependencies.wasm-bindgen]
-version = "^0.2"
 optional = true
+version = "^0.2"
 
 [dependencies.wasm-bindgen-futures]
-version = "^0.3"
 optional = true
-
-[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-reqwest = "^0.9"
+version = "^0.4"
 
 [dev-dependencies]
+tokio = {version = "^0.2", features = ["rt-threaded", "macros"]}
 # Note: If we bumpup wasm-bindge-test version, we should change CI setting.
-wasm-bindgen-test = "^0.2"
+wasm-bindgen-test = "^0.3"
 
 [features]
-web = [
-    "anyhow",
-    "thiserror",
-    "futures",
-    "js-sys",
-    "log",
-    "wasm-bindgen",
-    "wasm-bindgen-futures",
-    "web-sys",
-]
+client = ["reqwest"]
+default = ["client"]
+web = []
